chore: Moved rootLayout
